added optimized geometry writing into a new input file

+++ GaussianScripts/GaussianInputGen.py
@@ -2,27 +2,18 @@ import os
 import csv
 import numpy as np
 import pandas as pd
+from GaussianAnalyse import readOutput, readOptimizedGeom
 
-# This function generates a list of gaussian input names and titles as well as extracts
-# the names of original pdb files and their relative paths.
-# Rootdir argument is the directory from which the directory tree search starts
-
-def generateInputNames(rootdir='.', file_prefix='run', readPDB=True, readCOM=False):
+def generateInputNamesFromPDB(rootdir='.', file_prefix='run'):
+    """This function generates a list of gaussian input names, titles 
+    and names from the original pdb files and their relative paths.
+    Rootdir argument is the directory from which the directory tree search starts."""
 
     rootdir_length = len(rootdir) + 1 # length of the root directory plus '/' sign
 
     gaussian_titles = []
     gaussian_input_names = []
-
-    assert readPDB !=
-
-    if readPDB:
-        pdb_filenames = []
-        readCOM = False
-    
-    if readCOM:
-        com_filenames = []
-
+    pdb_filenames = []
     
     dir_counter = -1 # do not include the root dir
     file_counter = 0
@@ -45,10 +36,10 @@ def generateInputNames(rootdir='.', file_prefix='run', readPDB=True, readCOM=Fal
     
     return(pdb_filenames, gaussian_input_names, gaussian_titles)
 
-# This function reads a single PDB file and extracts all the useful data
-# for the gaussian input file
 
 def readPDB(filename):
+    """This function reads a single PDB file and extracts all the useful data
+    for the gaussian input file"""
 
     atom_data = []
 
@@ -68,11 +59,11 @@ def readPDB(filename):
     
     return(atom_data)
 
-# This function writes a single gaussian input file
-# based on the data from the PDB file and
-# some additional gaussian-related info
 
 def writeGaussianInput(filename, title, atom_data, ncores, method, basis_set, keywords, charge, multiplicity):
+    """This function writes a single gaussian input file
+    based on the data from the PDB file and
+    some additional gaussian-related info."""
     
     with open(filename, 'w') as file:
         file.write(f'%nprocshared={ncores} \n')
@@ -85,12 +76,12 @@ def writeGaussianInput(filename, title, atom_data, ncores, method, basis_set, ke
         
     return
 
-# This function generates gaussian input files corresponding
-# to every PDB file that is found in the directory tree
 
 def generateGaussianInput(pdb_filenames, gaussian_input_names, gaussian_titles, 
                           ncores='6', method='b3pw91', basis_set="6-31g(d')", 
                           keywords='empiricaldispersion=gd3', charge='2', multiplicity='1'):
+    """This function generates gaussian input files corresponding
+    to every PDB file that is found in the directory tree"""
 
     if len(pdb_filenames) > 0:
     
@@ -107,26 +98,28 @@ def generateGaussianInput(pdb_filenames, gaussian_input_names, gaussian_titles,
         print('No pdb files were found! \n')
 
 
-# Generates N geometry optimization input files with the lowest energy.
-# The function uses energy stats csv file from GaussianAnalyse.py
-# It is implied that for every single point energy calculation output file 'run.log'
-# there is a 'run.com' input file. The function generates the geometry optimization
-# input file with the same geometry but with added keyword 'opt'
+def generateGaussianOpt(energy_stats_filename='energy.csv', conformations_to_optimize=1, opt_keywords='opt', file_prefix='run_opt', keep_geometry=True):
+    """Generates N geometry optimization input files with the lowest energy.
+    The function uses energy stats csv file from GaussianAnalyse.py.
+    It is implied that for every single point energy calculation output file 'run.log'
+    there is a 'run.com' input file. The function generates the geometry optimization
+    input file with the same geometry but with added keyword 'opt'.
+    
+    keep_geometry=True: keep the same geometry as in the original input file
 
-def generateGaussianOpt(energy_stats_filename='energy.csv', conformations_to_optimize=1, opt_keywords='opt', file_prefix='run_opt'):
+    keep_geometry=False: read optimized geometry from the output file, write
+    it into the new input file
+    """
 
     print(f'Reading {energy_stats_filename}... \n')
 
     energy_stats = pd.read_csv(energy_stats_filename, delimiter=' ').sort_values('SCF energy, Hartree')
-
     selected_output_filenames = energy_stats.iloc[:, 0][:conformations_to_optimize].tolist()
-
     assert selected_output_filenames != [], 'Could not find any confirmations to optimize!'
 
     print(f'Generating geometry optimization files of {conformations_to_optimize} conformations with the lowest energy... \n')
 
     for i in range(len(selected_output_filenames)):
-
         output_filename = selected_output_filenames[i]
         original_input = output_filename.replace('.log', '.com')
         opt_filename = f'{file_prefix}-{i + 1}.com'
@@ -134,18 +127,54 @@ def generateGaussianOpt(energy_stats_filename='energy.csv', conformations_to_opt
         try:
             with open(original_input, 'r') as file_input:
                 lines = file_input.readlines()
+                # Read charge and multiplicity data:
+                for line in lines:
+                    numbers = 0
+                    for element in line.split():
+                        if element.replace('-', '').isnumeric():
+                            numbers += 1
+                                if numbers == 2:
+                                    charge_multiplicity_data = line
+
         except NameError:
             print(f'File {original_input} does not exist')
-        
-        with open(opt_filename, 'w') as file_opt:
-            for line in lines:
-                if '#' in line:
-                    file_opt.write(line.replace('\n', '') + opt_keywords + '\n')
-                elif 'Title' in line:
-                    file_opt.write(line.replace('\n', '') + 'optimization' + '\n')
-                else:
-                    file_opt.write(line)
 
+        if keep_geometry:
+            with open(opt_filename, 'w') as file_opt:
+                for line in lines:
+                    if '#' in line:
+                        file_opt.write(line.replace('\n', '') + opt_keywords + '\n')
+                    elif 'Title' in line:
+                        file_opt.write(line.replace('\n', '') + 'optimization' + '\n')
+                    else:
+                        file_opt.write(line)
+        else:
+            optimized_geometry_file = readOutput(output_filename)
+            optimized_geometry = readOptimizedGeom(optimized_geometry_file)
+
+            percent_section_written = False
+            hashtag_section_written = False
+            title_section_written = False
+            charge_multiplicity_section_written = False
+
+            with open(opt_filename, 'w') as file_opt:
+                for line in lines:
+                    if '%' in line:
+                        file_opt.write(line)
+                        percent_section_written = True
+                    elif '#' in line:
+                        file_opt.write(line.replace('\n', '') + opt_keywords + '\n')
+                        hashtag_section_written = True
+                    elif 'Title' in line:
+                        file_opt.write(line.replace('\n', '') + 'optimization' + '\n')
+                        title_section_written = True
+                    elif (percent_section_written and hashtag_section_written and title_section_written):
+                        file_opt.write(charge_multiplicity_data)
+                        charge_multiplicity_section_written = True
+                    elif (percent_section_written and hashtag_section_written and title_section_written and charge_multiplicity_section_written):
+                        for geometry_line in optimized_geometry:
+                            file_opt.write(geometry line + '\n')
+                    
     print('Done! \n')
